Fix issue with using conftest via stdin
Fixes #20. The addition of different parsers in support of TOML broke
the stdin behaviour, because we won't know the file name to detect the
extension. This commit goes part way to fixing that, by defaulting the
parser to YAML parser. This has a couple of advantages:

* Lots of files are YAML or JSON files but without an explicit
extension. Defaulting to that parser should make sense in lots of cases.
* This also reverts to the previous behaviour for stdin, parsing it as
JSON or YAML.

This doesn't fix the issue if you want to parse a TOML file via stdin.
At least that's not a regression so happy to fix that in a followup.
Adding an explicit flag to override the parser would be generally useful
here anyhow.