Fix issue with repeated output
If you used the same name for multiple tests, ie. called them all deny,
you got repeated output.

The test commands needs some refactoring, to split out the utilities and
test them independently. For the moment to get this fix out quickly I've
just added an acceptance test which works but is a little too coupled to
the current output.