chore: Refactor Execute command after rebasing
The execute command branch had become quite outdated. At the time there was not an easy way to output to the CLI in a similar manner as the test command, which is the case with the new OutputManager (although it would make sense to move it out of the test package). This reworks the execute command to apply the latest changes in the project.

Some notes: It does not work well with json yet, as the data is added to the structure as multiple jsonCheckresults, which are not merged on filename.