Handle explicitly blank filenames passed into conftest
Currently when a blank filename is passed to conftest, for example with
`conftest test ""` we try and open blank and fail with an odd error.

This change means we ignore explicit blank files passed in.

This hit me with the GitHub Action, due to the variadic nature of
passing files to conftest. GitHub Actions doesn't have a good optional
argument pattern, and passes "" explicitly. This change hits an edge
case for typical usage, but opens up the ability for the Action to take
arbitrary arguments.