Add plugin system for conftest
Conftest plugins provide a mechanism to extend conftest without needing to add to the codebase. This is ideal for adding specialized functionality (e.g. a kubectl integration). Plugins can be downloaded using go-getter. Plugins are defined by a plugin.yaml file, which is stored in the conftest cache directory (by default $HOME/conftest).