misc: More github.com/instrumenta => github.com/open-policy-agent updates
Ran a quick grep to find other instances of
github.com/instrumenta/conftest and these seemed to be the important
ones to update. Note there are still references back to github.com/instrumenta in
but those should be fine. Also, the image repository is still
instrumenta/conftest but we can update those once CI is running.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>