Acceptance test for new example test case (#355)
Add a Rego test case in examples/ that contains a divide by zero.

OPA will flag this as an error when running tests with verify so the
acceptance test should also expect a non-zero return code from conftest.

The new acceptance test is currently failing because if the test is not
a Fail then it defaults to a pass.

Signed-off-by: Stephen Kershaw <stekershaw@gmail.com>