fix: Flags with hyphens were skipped with config from ENV
Previously when a viper.GetX call was made for a flag with a hypen in
it, if that was set via the environment it would fail because viper's
internal store had it as a value with an underscore instead of a hyphen.
By adding this replacer the conversion happens automatically for
configuration in the environment without affecting the existing flags.

Signed-off-by: James Alseth <james@jalseth.me>