fix: Incorrect hash in go.sum
Dependabot introduced [1] an incorrect hash in the go.sum file for the
github.com/google/flatbuffers@v1.12.0 dependency. After working with
Github support, the root cause was determined to be that the default
public Go modules proxy has a different hash than if you connect
directly, likely because at some point the tags were swapped. Dependabot
runs with GOPRIVATE=* set, so it does not use the default proxy and is
why the hash is different.

This change uses the same hash as the public Go modules proxy in order
to ensure the builds work and anyone consuming the Conftest repository
are able to build. The flatbuffers maintainers introduced a 1.12.1 which
is the same code as the 1.12.0 release to fix [2] this issue going forward,
which we will use once the dependency that introduced the 1.12.0 version
is updated.

[1] https://github.com/open-policy-agent/conftest/pull/607
[2] https://github.com/google/flatbuffers/issues/6466

Signed-off-by: James Alseth <james@jalseth.me>