Support for terragrunt default config files
As stated in the following url,

https://terragrunt.gruntwork.io/docs/getting-started/configuration/

There are two known default filenames which terragrunt will look for
when running; terragrunt.hcl and terragrunt.hcl.json.

The latter of these files is presently handled by conftest. The former,
is incorrectly detected due to the hcl file extension. Terragrunt uses
HCL2 as of v0.19 when it moved to the terragrunt.hcl file format (and
away from the terraform.tfvars which it used to use).

https://github.com/gruntwork-io/terragrunt/blob/master/docs/_docs/06_migration_guides/upgrading_to_terragrunt_0.19.x.md

This patch adds specific detection of the known terragrunt configuration
files and sets their type to HCL2

Signed-off-by: Tim Rupp <caphrim007@gmail.com>