Refactor JUnit output formatting (#699)
To better integrate with CI systems such as Jenkis, the JUnit output now
includes the test namespace in the package name using dot notation. This
enables the tests under a namespace to be grouped where previously all
tests were under a single "conftest" group.

A `--junit-hide-message` flag was introduced to optionally suppress the
test summary message from the test name. This can be useful for tracking
trends over time since all violations for a file in a given namespace
will have a consistent test name.

Signed-off-by: James Alseth <james@jalseth.me>