feat: upgrade to ORAS Go v2
This removes the ORAS Go v1 dependency and upgrades to ORAS Go v2. The
support for authentication via docker dependency was removed in ORAS Go
v2 so parts of that needed to be replaced with cpuguy83/dockercfg[1].

Tests were expanded to include a TLS secured and basic auth protected
registry.

Signed-off-by: Zoran Regvart <zoran@regvart.com>

[1] github.com/cpuguy83/dockercfg

Signed-off-by: Zoran Regvart <zoran@regvart.com>