fix: examples/compose - a deny rule was incorrectly implemented
The conversion to a number of the version string evaluated to a boolean.
Subsequently the boolean was compared to a number, which always resulted in
a failure to comply to that rule.

This PR fixes the conversion error so that the comparison now works as expected.
This PR also adds some tests to prevent regression.

Signed-off-by: Dennis Geurts <github@dennisg.nl>