feat: Use Regal for Rego linting
This PR introduces [Regal](https://github.com/styrainc/regal) for linting
the example policies in this project. The Rego found there is generally in
a good shape, and the fixes done as part of this are mostly related to style.

The code has been updated to conformance against the following rules:

* [line-length](https://docs.styra.com/regal/rules/style/line-length)
* [use-assignment-operator](https://docs.styra.com/regal/rules/style/use-assignment-operator)
* [use-some-for-output-vars](https://docs.styra.com/regal/rules/idiomatic/use-some-for-output-vars)
* [prefer-snake-case](https://docs.styra.com/regal/rules/style/prefer-snake-case)
* [print-or-trace-call](https://docs.styra.com/regal/rules/testing/print-or-trace-call)

A few rules are also ignored for now (`.regal/config.yaml`) and I'll leave
it up to the maintainers to decide whether more rules should be enabled
later.

The PR also adds Regal to CI to lint any Rego introduced or changed
in future PRs.

For more information about Regal and its rules, see the
[documentation](https://docs.styra.com/regal) for the project.

Signed-off-by: Anders Eknert <anders@styra.com>