feat: Use ORAS credentials
The OCI go-getter support implemented in Conftest uses
github.com/cpuguy83/dockercfg. With ORAS v2.4.0 the credentials
support[1] was integrated in the oras-go. This makes for a good
opportunity to switch to using that instead. In addition to using the
supported credentials implementation this allows the use of Docker
native credential helpers[2].

This implementation is also a bit more correct, as it avoids using the
global default HTTP client, so mutations of it, e.g. when setting the
user-agent, do not cause race issues.

[1] https://github.com/oras-project/oras-credentials-go
[2] https://docs.docker.com/engine/reference/commandline/login/#credential-helpers

Signed-off-by: Zoran Regvart <zoran@regvart.com>