fix: Switch back to original --update behavior
In conftest v0.61.0 the behavior of --update changed in order to fetch
the policy URLs and stored them in a temporary directory.

It broke the previous behavior that updated current policy directory
and instead aggregated the current policy (under `policy`) with the
one(s) used in `--update`.

It also broke the behavior of a project that had no existing `policy`.

This reverts commit ed0ff0a76feacc1707b0c3b23e533a464c988cd0.

Fixes #1136.

Signed-off-by: Leonardo Taccari <leot@NetBSD.org>